+++ config/config.exs
@@ -57,8 +57,7 @@ config :td_cache, :event_stream,
   consumer_id: "default",
   consumer_group: "dd",
   streams: [
-    [key: "data_structure:events", consumer: TdDd.Cache.StructureLoader],
-    [key: "data_field:events", consumer: TdDd.Cache.FieldLoader]
+    [key: "data_structure:events", consumer: TdDd.Cache.StructureLoader]
   ]
 
 import_config "metadata.exs"