+++ lib/td_dd/data_structures/data_structure_version.ex
@@ -2,7 +2,6 @@ defmodule TdDd.DataStructures.DataStructureVersion do
   @moduledoc false
   use Ecto.Schema
   import Ecto.Changeset
-  alias TdDd.DataStructures.DataField
   alias TdDd.DataStructures.DataStructure
   alias TdDd.DataStructures.DataStructureRelation
   alias TdDd.DataStructures.DataStructureVersion
@@ -13,7 +12,6 @@ defmodule TdDd.DataStructures.DataStructureVersion do
   schema "data_structure_versions" do
     field(:version, :integer, default: 0)
     belongs_to(:data_structure, DataStructure)
-    many_to_many(:data_fields, DataField, join_through: "versions_fields", on_delete: :delete_all)
 
     many_to_many(:children, DataStructureVersion,
       join_through: DataStructureRelation,
@@ -46,17 +44,12 @@ defmodule TdDd.DataStructures.DataStructureVersion do
 
   def search_fields(%DataStructureVersion{
         data_structure: structure,
-        data_fields: fields,
         version: version
       }) do
-    fields_search_fields =
-      fields
-      |> Enum.map(&DataField.search_fields/1)
 
     structure
     |> DataStructure.search_fields()
     |> Map.put(:version, version)
-    |> Map.put(:data_fields, fields_search_fields)
   end
 
   def index_name(_) do